PE-82 | Integrated Video Communicatin in Chat screen
Chat screen displays video call system messages in the middle of the chat screen with an answer button for the call receiver.
Renamed "chat_sceen" to "chat_screen". Removed video communication demo screen.
